+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlClientIntegrationTests.cs
@@ -23,17 +23,15 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
     }
 
     [EnabledOnDockerPlatformTheory(EnabledOnDockerPlatformTheoryAttribute.DockerPlatform.Linux)]
-    [InlineData(CommandType.Text, "select 1/1", false)]
-    [InlineData(CommandType.Text, "select 1/1", false, true)]
-    [InlineData(CommandType.Text, "select 1/0", false, false, true)]
-    [InlineData(CommandType.Text, "select 1/0", false, false, true, false, false)]
-    [InlineData(CommandType.Text, "select 1/0", false, false, true, true, false)]
-    [InlineData(CommandType.StoredProcedure, "sp_who", false)]
-    [InlineData(CommandType.StoredProcedure, "sp_who", true)]
+    [InlineData(CommandType.Text, "select 1/1")]
+    [InlineData(CommandType.Text, "select 1/1", true)]
+    [InlineData(CommandType.Text, "select 1/0", false, true)]
+    [InlineData(CommandType.Text, "select 1/0", false, true, false, false)]
+    [InlineData(CommandType.Text, "select 1/0", false, true, true, false)]
+    [InlineData(CommandType.StoredProcedure, "sp_who")]
     public void SuccessfulCommandTest(
         CommandType commandType,
         string commandText,
-        bool captureStoredProcedureCommandName,
         bool captureTextCommandContent = false,
         bool isFailure = false,
         bool recordException = false,
@@ -52,12 +50,7 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
             .AddInMemoryExporter(activities)
             .AddSqlClientInstrumentation(options =>
             {
-#if !NETFRAMEWORK
-                options.SetDbStatementForStoredProcedure = captureStoredProcedureCommandName;
                 options.SetDbStatementForText = captureTextCommandContent;
-#else
-                options.SetDbStatementForText = captureStoredProcedureCommandName || captureTextCommandContent;
-#endif
                 options.RecordException = recordException;
                 if (shouldEnrich)
                 {
@@ -91,8 +84,25 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
         Assert.Single(activities);
         var activity = activities[0];
 
-        SqlClientTests.VerifyActivityData(commandType, commandText, captureStoredProcedureCommandName, captureTextCommandContent, isFailure, recordException, shouldEnrich, activity);
+        SqlClientTests.VerifyActivityData(commandType, commandText, captureTextCommandContent, isFailure, recordException, shouldEnrich, activity);
         SqlClientTests.VerifySamplingParameters(sampler.LatestSamplingParameters);
+
+        if (isFailure)
+        {
+#if NET
+            var status = activity.GetStatus();
+            Assert.Equal(ActivityStatusCode.Error, activity.Status);
+            Assert.Equal("Divide by zero error encountered.", activity.StatusDescription);
+            Assert.EndsWith("SqlException", activity.GetTagValue(SemanticConventions.AttributeErrorType) as string);
+            Assert.Equal("8134", activity.GetTagValue(SemanticConventions.AttributeDbResponseStatusCode));
+#else
+            var status = activity.GetStatus();
+            Assert.Equal(ActivityStatusCode.Error, activity.Status);
+            Assert.Equal("8134", activity.StatusDescription);
+            Assert.EndsWith("SqlException", activity.GetTagValue(SemanticConventions.AttributeErrorType) as string);
+            Assert.Equal("8134", activity.GetTagValue(SemanticConventions.AttributeDbResponseStatusCode));
+#endif
+        }
     }
 
     private string GetConnectionString()